Add static_cast to height and width in func call in main
width and height are "qreal" (typedef for double). Having them
implicitly cast to int leads to warnings on some compilers, and can lead
to doubt whether they're supposed to be converted.

+++ src/main.cpp
@@ -420,8 +420,10 @@ int main( int argc, char* argv[] )
             m_pWindow->setGeometry(
                 0,
                 0,
-                qobject_cast<QQuickItem*>( quickObj )->width(),
-                qobject_cast<QQuickItem*>( quickObj )->height() );
+                static_cast<int>(
+                    qobject_cast<QQuickItem*>( quickObj )->width() ),
+                static_cast<int>(
+                    qobject_cast<QQuickItem*>( quickObj )->height() ) );
             m_pWindow->show();
         }
